pyproject.toml
[tool.poetry]
name = "starknet_devnet"
version = "0.2.10"
description = "A local testnet for Starknet"
authors = ["FabijanC <fabijan.corak@gmail.com>"]
license = "ISC"

readme = "README.md"
repository = "https://github.com/Shard-Labs/starknet-devnet"
homepage = "https://github.com/Shard-Labs/starknet-devnet"
keywords = ["starknet", "cairo", "testnet", "local", "server"]

[tool.poetry.dependencies]
python = ">=3.7.2,<3.10"
Flask = {extras = ["async"], version = "~2.0.3"}
flask-cors = "~3.0.10"
cairo-lang = "0.9.1"
Werkzeug = "~2.0.3"
cloudpickle = "~2.1.0"
crypto-cpp-py = "~1.0.4"
marshmallow = "~3.17.0"
typing-extensions = "~4.3.0"
gunicorn = "~20.1.0"

[tool.poetry.dev-dependencies]
pylint = "~2.12.2"
web3 = "~5.28.0"
psutil = "~5.9.1"
jsonschema = "~3.2.0"
pytest-xdist = "~2.5.0"
pylint-quotes = "~0.2.3"
black = "~22.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
starknet-devnet = "starknet_devnet.server:main"

[tool.pytest.ini_options]
markers = [
    "account",
    "account_predeployed",
    "call",
    "declare",
    "deploy",
    "estimate_fee",
    "fee_token",
    "general_workflow",
    "invoke",
    "restart",
    "state_update",
    "timestamps",
    "transaction_trace",
    "tx_version",
    "web3_messaging"
]
junit_family="xunit1"
asyncio_mode="strict"
